# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: EasyWords API
  description: |
    API для платформы изучения словарного запаса EasyWords.

    Пользователи могут сохранять слова, изучать их с помощью флеш-карточек,
    получать примеры использования слов и делиться ими с другими пользователями.

    API соответствует спецификации JSON:API.
  version: 1.0.0
  contact:
    name: EasyWords Team
    url: https://easywordsapp.ru
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.easywordsapp.ru/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /words:
    get:
      tags:
        - Words
      summary: Получить список слов пользователя
      description: |
        Получает список слов авторизованного пользователя с учётом настроек:
        - фильтрация по языку (обязательный параметр)
        - режим сортировки
        - показ/скрытие изученных, импортированных и общих слов
      operationId: getUserWords
      security:
        - BearerAuth: []
      parameters:
        - name: language
          in: query
          description: Код языка для фильтрации (EN, DE, ES, FR, IT, RU)
          required: true
          schema:
            type: string
            enum: [EN, DE, ES, FR, IT, RU]
            example: EN
      responses:
        '200':
          description: Список слов пользователя
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordsPaginatedResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    post:
      tags:
        - Words
      summary: Создать новое слово
      description: |
        Создает новое слово для авторизованного пользователя.
        Принимает оригинальное слово, перевод и код языка.
      operationId: createWord
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              type: object
              required:
                - original
                - translated
                - language
              properties:
                original:
                  type: string
                  maxLength: 255
                  description: Оригинальное слово
                  example: hello
                translated:
                  type: string
                  maxLength: 255
                  description: Перевод слова
                  example: привет
                language:
                  type: string
                  maxLength: 10
                  description: Код языка
                  enum: [EN, DE, ES, FR, IT, RU]
                  example: EN
      responses:
        '201':
          description: Слово успешно создано
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /words/with-filters:
    get:
      tags:
        - Words
      summary: Получить список слов с фильтрацией
      description: |
        Получает список слов авторизованного пользователя с гибкой фильтрацией.
        Поддерживает фильтрацию по статусу изучения, общим словам,
        источнику и избранным словам.
      operationId: getUserWordsWithFilters
      security:
        - BearerAuth: []
      parameters:
        - name: done
          in: query
          description: Фильтрация по статусу изучения
          required: false
          schema:
            type: string
            enum: [true, false]
        - name: shared
          in: query
          description: Фильтрация по общим словам
          required: false
          schema:
            type: string
            enum: [true, false]
        - name: from_sample
          in: query
          description: Фильтрация по источнику (импортированные)
          required: false
          schema:
            type: string
            enum: [true, false]
        - name: starred
          in: query
          description: Фильтрация по избранным
          required: false
          schema:
            type: string
            enum: [true, false]
      responses:
        '200':
          description: Список слов с фильтрацией
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordsPaginatedResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /random:
    get:
      tags:
        - Words
      summary: Получить случайные слова
      description: |
        Получает случайные слова для изучения с учётом настроек пользователя:
        - фильтрация по языку (default_language)
        - исключение изученных слов (если known_enabled === false)
        - ограничение количества слов (по умолчанию 20)
      operationId: getRandomWords
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Максимальное количество слов
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Массив случайных слов
          content:
            application/vnd.api+json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WordResource'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /words/{wordId}:
    get:
      tags:
        - Words
      summary: Получить слово по ID
      description: |
        Возвращает слово по ID. Проверяет, что слово принадлежит авторизованному пользователю.
      operationId: getWord
      security:
        - BearerAuth: []
      parameters:
        - name: wordId
          in: path
          description: ID слова
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные слова
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Слово не найдено
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Words
      summary: Удалить слово
      description: |
        Удаляет слово по ID. Проверяет, что слово принадлежит авторизованному пользователю.
      operationId: deleteWord
      security:
        - BearerAuth: []
      parameters:
        - name: wordId
          in: path
          description: ID слова
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Слово успешно удалено
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Слово не найдено
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /words/{wordId}/views:
    post:
      tags:
        - Words
      summary: Увеличить счётчик просмотров
      description: |
        Увеличивает счётчик просмотров слова на 1.
      operationId: incrementWordViews
      security:
        - BearerAuth: []
      parameters:
        - name: wordId
          in: path
          description: ID слова
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Счётчик просмотров увеличен
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /words/{wordId}/starred:
    put:
      tags:
        - Words
      summary: Изменить статус избранного
      description: |
        Обновляет статус избранного у слова (true или false).
      operationId: toggleWordStarred
      security:
        - BearerAuth: []
      parameters:
        - name: wordId
          in: path
          description: ID слова
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              type: object
              required:
                - starred
              properties:
                starred:
                  type: boolean
                  description: Статус избранного
                  example: true
      responses:
        '200':
          description: Статус избранного обновлен
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /words/{wordId}/learned:
    post:
      tags:
        - Words
      summary: Пометить слово как выученное
      description: |
        Помечает слово как выученное, устанавливая поле done_at в текущее время.
      operationId: markWordAsLearned
      security:
        - BearerAuth: []
      parameters:
        - name: wordId
          in: path
          description: ID слова
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Слово помечено как выученное
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /words/share:
    post:
      tags:
        - Words
      summary: Поделиться словом с другим пользователем
      description: |
        Создает копию слова для указанного пользователя на основе оригинального слова.
        Проверяет, что оригинальное слово принадлежит авторизованному пользователю.
      operationId: shareWord
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              type: object
              required:
                - word_id
                - user_id
              properties:
                word_id:
                  type: integer
                  description: ID слова для шаринга
                  example: 123
                user_id:
                  type: integer
                  description: ID пользователя, которому отправляется слово
                  example: 456
      responses:
        '201':
          description: Слово успешно отправлено
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/WordResponse'
        '401':
          description: Пользователь не авторизован
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Слово не принадлежит пользователю
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Слово или пользователь не найдены
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации (Laravel Sanctum)
  schemas:
    WordResource:
      type: object
      required:
        - type
        - id
        - attributes
      properties:
        type:
          type: string
          enum: [words]
          example: words
        id:
          type: string
          example: "123"
        attributes:
          type: object
          required:
            - original
            - translated
            - language
            - starred
            - views
            - from_sample
            - created_at
            - updated_at
          properties:
            original:
              type: string
              description: Оригинальное слово
              example: hello
            translated:
              type: string
              description: Перевод слова
              example: привет
            language:
              type: string
              description: Код языка
              enum: [EN, DE, ES, FR, IT, RU]
              example: EN
            done_at:
              type: string
              format: date-time
              nullable: true
              description: Дата и время пометки слова как выученного
              example: null
            starred:
              type: boolean
              description: Статус избранного
              example: false
            views:
              type: integer
              description: Количество просмотров слова
              example: 5
            from_sample:
              type: boolean
              description: Импортировано ли из samples
              example: false
            created_at:
              type: string
              format: date-time
              description: Дата и время создания
              example: "2026-01-11T12:00:00Z"
            updated_at:
              type: string
              format: date-time
              description: Дата и время последнего обновления
              example: "2026-01-11T12:00:00Z"
    WordResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/WordResource'
    WordsPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WordResource'
        meta:
          type: object
          properties:
            current_page:
              type: integer
              example: 1
            per_page:
              type: integer
              example: 20
            total:
              type: integer
              example: 100
            last_page:
              type: integer
              example: 5
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
                example: "401"
              title:
                type: string
                example: Unauthenticated
              detail:
                type: string
                nullable: true
                example: Word not found or does not belong to user
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.
            password:
              - The password field is required.